Refactor server fact loading into a separate class
puppet 7.x still uses legacy facts with symbolic fact names, e.g.
value(:hostname)

(cherry picked from commit 8068bcec0ed0c8386e66a1e202648bbe8c51a52c)
